feat: 아티클 등록 아웃박스 엔티티와 저장소 명칭을 정리한다
