+++ infrastructure/jpa/src/main/kotlin/site/techmoa/infrastructure/jpa/repository/ArticleCreatedOutboxJpaRepository.kt
@@ -1,0 +1,7 @@
+package site.techmoa.infrastructure.jpa.repository
+
+import org.springframework.data.jpa.repository.JpaRepository
+import site.techmoa.infrastructure.jpa.entity.ArticleCreatedOutboxEntity
+
+interface ArticleCreatedOutboxJpaRepository : JpaRepository<ArticleCreatedOutboxEntity, Long> {
+}
